requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.7
django-elasticsearch-dsl==8.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
requests
pandas
